src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ollama>=0.6.1",
    "openai>=2.8.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uuid>=1.30",
]
